requirements.txt
langchain
wikipedia
langchain-community
python-dotenv
pydantic
langchain-mistralai
flask
python-docx
